pyproject.toml
[project]
name = "qa-documents-modeling"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Isaias Gutierrez Cruz", email = "isaiasgutierrezqfb@gmail.com" }
]
requires-python = ">=3.10, <4.0"
dependencies = [
    "dotenv>=0.9.9",
    "llama-index-core>=0.12.27",
    "llama-index-llms-openai>=0.3.29",
    "llama-index-node-parser-docling>=0.3.1",
    "llama-index-readers-docling>=0.3.2",
    "nest-asyncio>=1.6.0",
    "sentence-transformers>=4.0.1",
    "streamlit>=1.44.0",
    "llama-index-embeddings-huggingface>=0.5.2",
    "sf-hamilton>=1.88.0",
    "llama-index-finetuning>=0.3.1",
    "datasets>=3.5.0",
    "huggingface-hub>=0.30.1",
    "ipywidgets>=8.1.5",
    "transformers>=4.50.3",
    "peft>=0.15.1",
    "ragas>=0.2.14",
    "annoy>=1.17.3",
]

[project.scripts]
qa-documents-modeling = "qa_documents_modeling:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project.optional-dependencies]
lint = [
    "ruff>=0.11.2",
]
dev = [
    "jupyterlab>=4.3.6",
]

[tool.ruff]
line-length = 90
fix = true

[tool.ruff.lint]
preview = true
explicit-preview-rules = true

select = [
  "ALL",
]
ignore = [
  "TD003",
  "TD004",
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["S101", "D103"]

[tool.ruff.lint.pydocstyle]
convention = "numpy"

[tool.ruff.format]
docstring-code-format = true
